+++ next.config.ts
@@ -2,12 +2,14 @@ import { withPayload } from '@payloadcms/next/withPayload'
 import type { NextConfig } from 'next'
 
 const nextConfig: NextConfig = {
+	reactCompiler: true,
 	reactStrictMode: true,
 	typescript: {
 		ignoreBuildErrors: true,
 	},
 	images: {
 		formats: ['image/avif'],
+		qualities: [90],
 		dangerouslyAllowSVG: true,
 		minimumCacheTTL: 60 * 60 * 24 * 30,
 		remotePatterns: [
@@ -17,16 +19,26 @@ const nextConfig: NextConfig = {
 			},
 		],
 	},
+	logging: {
+		fetches: {
+			fullUrl: true,
+		},
+	},
 	experimental: {
 		useCache: true,
-		reactCompiler: false,
+		turbopackFileSystemCacheForDev: true,
+		optimizeCss: true,
+		optimizeServerReact: true,
 		optimizePackageImports: [
 			'@mantine/core',
 			'@mantine/hooks',
 			'@mantine/form',
-			'@mantine/carousel',
 			'@mantine/nprogress',
+			'@mantine/notifications',
 		],
+		serverActions: {
+			bodySizeLimit: '10mb',
+		},
 	},
 }