+++ src/admin/collections/Pages.ts
@@ -29,6 +29,7 @@ export const Pages: CollectionConfig = {
 	admin: {
 		useAsTitle: 'title',
 		defaultColumns: ['title', 'slug', '_status', 'updatedAt', 'author'],
+		group: 'Content',
 		livePreview: {
 			url: ({ data, req }) => {
 				return generatePreviewPath({