+++ src/lib/server-functions/loader.ts
@@ -1,5 +1,5 @@
 'use server'
-import { unstable_cacheTag as cacheTag } from 'next/cache'
+import { cacheTag } from 'next/cache'
 import { draftMode } from 'next/headers'
 import { getPayload } from 'payload'