+++ src/proxy.ts
@@ -3,19 +3,12 @@ import { NextRequest, NextResponse } from 'next/server'
 
 import { slugReusable } from '$modules/vars'
 
-export async function middleware(request: NextRequest) {
+export async function proxy(request: NextRequest) {
 	const { pathname } = request.nextUrl
 
 	const response = NextResponse.next()
 	const cookieStore = await cookies()
 	const isLoggedIn = cookieStore.has('payload-token')
-
-	// Redirect to 404 when visitor not loggedin and trying to access /uploads
-	if (pathname.startsWith('/uploads')) {
-		if (!isLoggedIn) {
-			return NextResponse.redirect(new URL('/404', request.url))
-		}
-	}
 
 	if (pathname.startsWith(`/${slugReusable}`) && !isLoggedIn) {
 		return NextResponse.redirect(new URL('/404', request.url))